moodbar: check QSlider max/min before using them in division
When switching playback from an item for which the moodbar is not
displayed (e.g. an internet stream), to an item for which it is
displayed (e.g. a local mp3 file), Clementine sometimes crashes.

This happens because the slider_opt->maximum and slider_opt->minimum
have the value 0 (their default value), and the difference is used as a
divisor.  This gives a division by 0, and a SIGFPE.

This problem has been fixed in commit 2741a7b.  However, when 5f3afdf
reverted 5353a3a, a little to much was reverted and we lost what 2741a7b
did.  This patch re-introduces the fix.

Fixes #5261

moodbar: Add comment